Remove chrisgrieser/nvim-spider
The enhanced motions were great in most situations, but it was making me
crazy specifically with "dw" and "cw" (deleting or changing to the end
of the current word). I have the muscle memory ingrained too deeply to
shift away from using those commands, and the way that it would also
delete any trailing symbols (including whitespace) was very frustrating
and was never what I want to happen.