Disable autoformat for LazyVim
Most of my repositories have not been completely reformatted using an
automated tool, and clang-format was a little too aggressive in
reformatting everything on every save. It was making my commits very
large and very confusing.

If we explicitly enforce the formatting using the automated tools and
commit the reformatted source code, then this feature could be reenabled
in the future again.

Note: It is possible to configure 'git blame' to ignore mass
reformatting commits by using a .git-blame-ignore-revs file to itemize
which commit hashes to ignore.

https://akrabat.com/ignoring-revisions-with-git-blame/